Optimize team stats with pre-computed possession metrics
- Fix team_season_stats schema issues (missing columns)
- Replace SQLite syntax (AS REAL) with PostgreSQL (AS NUMERIC)
- Add CORS support for localhost:3001
- Implement possession stats pre-computation (99% perf improvement)
- Add career stats aggregation from raw counts
- Create populate_possession_stats.py for batch processing
- Performance: 3-8s timeout → 0.5s response time

+++ backend/prompts.py
@@ -118,7 +118,7 @@ When asked about career statistics or stats "across all seasons":
      GROUP BY player_id
    )
    SELECT ct.full_name, ct.career_goals, gc.games_played,
-          ROUND(CAST(ct.career_goals AS REAL) / gc.games_played, 1) as goals_per_game
+          ROUND(CAST(ct.career_goals AS NUMERIC) / gc.games_played, 1) as goals_per_game
    FROM career_totals ct
    JOIN game_counts gc ON ct.player_id = gc.player_id
    WHERE gc.games_played > 0
